Replace load_imageset tests with xia2.setup tests
Directly testing the load_imageset function is fragile as there are a
number of singletons lurking in the depths of xia2 which can mean shared
state between tests. More reliable to test this behaviour in a separate
process via the xia2.setup command.
